Crate rusoto_batch

AWS Batch enables you to run batch computing workloads on the AWS Cloud. Batch computing is a common way for developers, scientists, and engineers to access large amounts of compute resources, and AWS Batch removes the undifferentiated heavy lifting of configuring and managing the required infrastructure. AWS Batch will be familiar to users of traditional batch computing software. This service can efficiently provision resources in response to jobs submitted in order to eliminate capacity constraints, reduce compute costs, and deliver results quickly.

As a fully managed service, AWS Batch enables developers, scientists, and engineers to run batch computing workloads of any scale. AWS Batch automatically provisions compute resources and optimizes the workload distribution based on the quantity and scale of the workloads. With AWS Batch, there is no need to install or manage batch computing software, which allows you to focus on analyzing results and solving problems. AWS Batch reduces operational complexities, saves time, and reduces costs, which makes it easy for developers, scientists, and engineers to run their batch jobs in the AWS Cloud.

If you're using the service, you're probably looking for BatchClient and Batch.

Structs

ArrayProperties
An object representing an AWS Batch array job.

ArrayPropertiesDetail
An object representing the array properties of a job.

ArrayPropertiesSummary
An object representing the array properties of a job.

AttemptContainerDetail
An object representing the details of a container that is part of a job attempt.

AttemptDetail
An object representing a job attempt.

BatchClient
A client for the AWS Batch API.

CancelJobRequest

CancelJobResponse

ComputeEnvironmentDetail
An object representing an AWS Batch compute environment.

ComputeEnvironmentOrder
The order in which compute environments are tried for job placement within a queue. Compute environments are tried in ascending order. For example, if two compute environments are associated with a job queue, the compute environment with a lower order integer value is tried for job placement first.

ComputeResource
An object representing an AWS Batch compute resource.

ComputeResourceUpdate
An object representing the attributes of a compute environment that can be updated.

ContainerDetail
An object representing the details of a container that is part of a job.

ContainerOverrides
The overrides that should be sent to a container.

ContainerProperties
Container properties are used in job definitions to describe the container that is launched as part of a job.

ContainerSummary
An object representing summary details of a container within a job.

CreateComputeEnvironmentRequest

CreateComputeEnvironmentResponse

CreateJobQueueRequest

CreateJobQueueResponse

DeleteComputeEnvironmentRequest

DeleteComputeEnvironmentResponse

DeleteJobQueueRequest

DeleteJobQueueResponse

DeregisterJobDefinitionRequest

DeregisterJobDefinitionResponse

DescribeComputeEnvironmentsRequest

DescribeComputeEnvironmentsResponse

DescribeJobDefinitionsRequest

DescribeJobDefinitionsResponse

DescribeJobQueuesRequest

DescribeJobQueuesResponse

DescribeJobsRequest

DescribeJobsResponse

Host
Determine whether your data volume persists on the host container instance and where it is stored. If this parameter is empty, then the Docker daemon assigns a host path for your data volume, but the data is not guaranteed to persist after the containers associated with it stop running.

JobDefinition
An object representing an AWS Batch job definition.

JobDependency
An object representing an AWS Batch job dependency.

JobDetail
An object representing an AWS Batch job.

JobQueueDetail
An object representing the details of an AWS Batch job queue.

JobSummary
An object representing summary details of a job.

JobTimeout
An object representing a job timeout configuration.

KeyValuePair
A key-value pair object.

LaunchTemplateSpecification
An object representing a launch template associated with a compute resource. You must specify either the launch template ID or launch template name in the request, but not both.

ListJobsRequest

ListJobsResponse

MountPoint
Details on a Docker volume mount point that is used in a job's container properties. This parameter maps to Volumes in the Create a container section of the Docker Remote API and the --volume option to docker run.

NetworkInterface
An object representing the elastic network interface for a multi-node parallel job node.

NodeDetails
An object representing the details of a multi-node parallel job node.

NodeOverrides
Object representing any node overrides to a job definition that is used in a SubmitJob API operation.

NodeProperties
An object representing the node properties of a multi-node parallel job.

NodePropertiesSummary
An object representing the properties of a node that is associated with a multi-node parallel job.

NodePropertyOverride
Object representing any node overrides to a job definition that is used in a SubmitJob API operation.

NodeRangeProperty
An object representing the properties of the node range for a multi-node parallel job.

RegisterJobDefinitionRequest

RegisterJobDefinitionResponse

ResourceRequirement
The type and amount of a resource to assign to a container. Currently, the only supported resource type is GPU.

RetryStrategy
The retry strategy associated with a job.

SubmitJobRequest

SubmitJobResponse

TerminateJobRequest

TerminateJobResponse

Ulimit
The ulimit settings to pass to the container.

UpdateComputeEnvironmentRequest

UpdateComputeEnvironmentResponse

UpdateJobQueueRequest

UpdateJobQueueResponse

Volume
A data volume used in a job's container properties.

Enums

CancelJobError
Errors returned by CancelJob

CreateComputeEnvironmentError
Errors returned by CreateComputeEnvironment

CreateJobQueueError
Errors returned by CreateJobQueue

DeleteComputeEnvironmentError
Errors returned by DeleteComputeEnvironment

DeleteJobQueueError
Errors returned by DeleteJobQueue

DeregisterJobDefinitionError
Errors returned by DeregisterJobDefinition

DescribeComputeEnvironmentsError
Errors returned by DescribeComputeEnvironments

DescribeJobDefinitionsError
Errors returned by DescribeJobDefinitions

DescribeJobQueuesError
Errors returned by DescribeJobQueues

DescribeJobsError
Errors returned by DescribeJobs

ListJobsError
Errors returned by ListJobs

RegisterJobDefinitionError
Errors returned by RegisterJobDefinition

SubmitJobError
Errors returned by SubmitJob

TerminateJobError
Errors returned by TerminateJob

UpdateComputeEnvironmentError
Errors returned by UpdateComputeEnvironment

UpdateJobQueueError
Errors returned by UpdateJobQueue

Traits

Batch
Trait representing the capabilities of the AWS Batch API. AWS Batch clients implement this trait.