Struct rusoto_batch::ComputeResource

pub struct ComputeResource {
    pub bid_percentage: Option<i64>,
    pub desiredv_cpus: Option<i64>,
    pub ec_2_key_pair: Option<String>,
    pub image_id: Option<String>,
    pub instance_role: String,
    pub instance_types: Vec<String>,
    pub maxv_cpus: i64,
    pub minv_cpus: i64,
    pub security_group_ids: Vec<String>,
    pub spot_iam_fleet_role: Option<String>,
    pub subnets: Vec<String>,
    pub tags: Option<HashMap<String, String>>,
    pub type_: String,
}

An object representing an AWS Batch compute resource.

Fields
bid_percentage: Option<i64>

The minimum percentage that a Spot Instance price must be when compared with the On-Demand price for that instance type before instances are launched. For example, if your bid percentage is 20%, then the Spot price must be below 20% of the current On-Demand price for that EC2 instance.

desiredv_cpus: Option<i64>

The desired number of EC2 vCPUS in the compute environment.

ec_2_key_pair: Option<String>

The EC2 key pair that is used for instances launched in the compute environment.

image_id: Option<String>

The Amazon Machine Image (AMI) ID used for instances launched in the compute environment.

instance_role: String

The Amazon ECS instance role applied to Amazon EC2 instances in a compute environment.

instance_types: Vec<String>

The instances types that may launched.

maxv_cpus: i64

The maximum number of EC2 vCPUs that an environment can reach.

minv_cpus: i64

The minimum number of EC2 vCPUs that an environment should maintain.

security_group_ids: Vec<String>

The EC2 security group that is associated with instances launched in the compute environment.

spot_iam_fleet_role: Option<String>

The Amazon Resource Name (ARN) of the Amazon EC2 Spot Fleet IAM role applied to a SPOT compute environment.

subnets: Vec<String>

The VPC subnets into which the compute resources are launched.

tags: Option<HashMap<String, String>>

Key-value pair tags to be applied to resources that are launched in the compute environment.

type_: String

The type of compute environment.