Struct rusoto_batch::ContainerProperties

pub struct ContainerProperties {
    pub command: Option<Vec<String>>,
    pub environment: Option<Vec<KeyValuePair>>,
    pub image: String,
    pub job_role_arn: Option<String>,
    pub memory: i64,
    pub mount_points: Option<Vec<MountPoint>>,
    pub privileged: Option<bool>,
    pub readonly_root_filesystem: Option<bool>,
    pub ulimits: Option<Vec<Ulimit>>,
    pub user: Option<String>,
    pub vcpus: i64,
    pub volumes: Option<Vec<Volume>>,
}

Container properties are used in job definitions to describe the container that is launched as part of a job.

Fields
command: Option<Vec<String>>

The command that is passed to the container. This parameter maps to Cmd in the Create a container section of the Docker Remote API and the COMMAND parameter to docker run. For more information, see https://docs.docker.com/engine/reference/builder/#cmd.

environment: Option<Vec<KeyValuePair>>

The environment variables to pass to a container. This parameter maps to Env in the Create a container section of the Docker Remote API and the --env option to docker run.

We do not recommend using plain text environment variables for sensitive information, such as credential data.

image: String

The image used to start a container. This string is passed directly to the Docker daemon. Images in the Docker Hub registry are available by default. Other repositories are specified with repository-url/image:tag . Up to 255 letters (uppercase and lowercase), numbers, hyphens, underscores, colons, periods, forward slashes, and number signs are allowed. This parameter maps to Image in the Create a container section of the Docker Remote API and the IMAGE parameter of docker run.

Images in Amazon ECR repositories use the full registry and repository URI (for example, 012345678910.dkr.ecr.<region-name>.amazonaws.com/<repository-name>).

Images in official repositories on Docker Hub use a single name (for example, ubuntu or mongo).

Images in other repositories on Docker Hub are qualified with an organization name (for example, amazon/amazon-ecs-agent).

Images in other online repositories are qualified further by a domain name (for example, quay.io/assemblyline/ubuntu).

job_role_arn: Option<String>

The Amazon Resource Name (ARN) of the IAM role that the container can assume for AWS permissions.

memory: i64

The hard limit (in MiB) of memory to present to the container. If your container attempts to exceed the memory specified here, the container is killed. This parameter maps to Memory in the Create a container section of the Docker Remote API and the --memory option to docker run.

mount_points: Option<Vec<MountPoint>>

The mount points for data volumes in your container. This parameter maps to Volumes in the Create a container section of the Docker Remote API and the --volume option to docker run.

privileged: Option<bool>

When this parameter is true, the container is given elevated privileges on the host container instance (similar to the root user). This parameter maps to Privileged in the Create a container section of the Docker Remote API and the --privileged option to docker run.

readonly_root_filesystem: Option<bool>

When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run.

ulimits: Option<Vec<Ulimit>>

A list of ulimits to set in the container. This parameter maps to Ulimits in the Create a container section of the Docker Remote API and the --ulimit option to docker run.

user: Option<String>

The user name to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run.

vcpus: i64

The number of vCPUs reserved for the container. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run. Each vCPU is equivalent to 1,024 CPU shares.

volumes: Option<Vec<Volume>>

A list of data volumes used in a job.