Struct rusoto_batch::CreateJobQueueRequest

pub struct CreateJobQueueRequest {
    pub compute_environment_order: Vec<ComputeEnvironmentOrder>,
    pub job_queue_name: String,
    pub priority: i64,
    pub state: Option<String>,
}

Fields
compute_environment_order: Vec<ComputeEnvironmentOrder>

The set of compute environments mapped to a job queue and their order relative to each other. The job scheduler uses this parameter to determine which compute environment should execute a given job. Compute environments must be in the VALID state before you can associate them with a job queue. You can associate up to 3 compute environments with a job queue.

job_queue_name: String

The name of the job queue.

priority: i64

The priority of the job queue. Job queues with a higher priority (or a lower integer value for the priority parameter) are evaluated first when associated with same compute environment. Priority is determined in ascending order, for example, a job queue with a priority value of 1 is given scheduling preference over a job queue with a priority value of 10.

state: Option<String>

The state of the job queue. If the job queue state is ENABLED, it is able to accept jobs.