Struct rusoto_batch::JobDefinition

pub struct JobDefinition {
    pub container_properties: Option<ContainerProperties>,
    pub job_definition_arn: String,
    pub job_definition_name: String,
    pub parameters: Option<HashMap<String, String>>,
    pub retry_strategy: Option<RetryStrategy>,
    pub revision: i64,
    pub status: Option<String>,
    pub timeout: Option<JobTimeout>,
    pub type_: String,
}

An object representing an AWS Batch job definition.

Fields
container_properties: Option<ContainerProperties>

An object with various properties specific to container-based jobs.

job_definition_arn: String

The Amazon Resource Name (ARN) for the job definition.

job_definition_name: String

The name of the job definition.

parameters: Option<HashMap<String, String>>

Default parameters or parameter substitution placeholders that are set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

retry_strategy: Option<RetryStrategy>

The retry strategy to use for failed jobs that are submitted with this job definition.

revision: i64

The revision of the job definition.

status: Option<String>

The status of the job definition.

timeout: Option<JobTimeout>

The timeout configuration for jobs that are submitted with this job definition. You can specify a timeout duration after which AWS Batch terminates your jobs if they have not finished.

type_: String

The type of job definition.