Struct rusoto_batch::BatchClient

pub struct BatchClient { /* fields omitted */ }

A client for the AWS Batch API.

Methods

impl BatchClient

pub fn new(region: Region) -> BatchClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> BatchClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,