Struct rusoto_batch::ComputeEnvironmentDetail

pub struct ComputeEnvironmentDetail {
    pub compute_environment_arn: String,
    pub compute_environment_name: String,
    pub compute_resources: Option<ComputeResource>,
    pub ecs_cluster_arn: String,
    pub service_role: Option<String>,
    pub state: Option<String>,
    pub status: Option<String>,
    pub status_reason: Option<String>,
    pub type_: Option<String>,
}

An object representing an AWS Batch compute environment.

Fields
compute_environment_arn: String

The Amazon Resource Name (ARN) of the compute environment.

compute_environment_name: String

The name of the compute environment.

compute_resources: Option<ComputeResource>

The compute resources defined for the compute environment.

ecs_cluster_arn: String

The Amazon Resource Name (ARN) of the underlying Amazon ECS cluster used by the compute environment.

service_role: Option<String>

The service role associated with the compute environment that allows AWS Batch to make calls to AWS API operations on your behalf.

state: Option<String>

The state of the compute environment. The valid values are ENABLED or DISABLED. An ENABLED state indicates that you can register instances with the compute environment and that the associated instances can accept jobs.

status: Option<String>

The current status of the compute environment (for example, CREATING or VALID).

status_reason: Option<String>

A short, human-readable string to provide additional details about the current status of the compute environment.

type_: Option<String>

The type of the compute environment.