Struct rusoto_batch::ContainerDetail

pub struct ContainerDetail {
    pub command: Option<Vec<String>>,
    pub container_instance_arn: Option<String>,
    pub environment: Option<Vec<KeyValuePair>>,
    pub exit_code: Option<i64>,
    pub image: Option<String>,
    pub job_role_arn: Option<String>,
    pub log_stream_name: Option<String>,
    pub memory: Option<i64>,
    pub mount_points: Option<Vec<MountPoint>>,
    pub privileged: Option<bool>,
    pub readonly_root_filesystem: Option<bool>,
    pub reason: Option<String>,
    pub task_arn: Option<String>,
    pub ulimits: Option<Vec<Ulimit>>,
    pub user: Option<String>,
    pub vcpus: Option<i64>,
    pub volumes: Option<Vec<Volume>>,
}

An object representing the details of a container that is part of a job.

Fields
command: Option<Vec<String>>

The command that is passed to the container.

container_instance_arn: Option<String>

The Amazon Resource Name (ARN) of the container instance on which the container is running.

environment: Option<Vec<KeyValuePair>>

The environment variables to pass to a container.

exit_code: Option<i64>

The exit code to return upon completion.

image: Option<String>

The image used to start the container.

job_role_arn: Option<String>

The Amazon Resource Name (ARN) associated with the job upon execution.

log_stream_name: Option<String> memory: Option<i64>

The number of MiB of memory reserved for the job.

mount_points: Option<Vec<MountPoint>>

The mount points for data volumes in your container.

privileged: Option<bool>

When this parameter is true, the container is given elevated privileges on the host container instance (similar to the root user).

readonly_root_filesystem: Option<bool>

When this parameter is true, the container is given read-only access to its root file system.

reason: Option<String>

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

task_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon ECS task that is associated with the container job.

ulimits: Option<Vec<Ulimit>>

A list of ulimit values to set in the container.

user: Option<String>

The user name to use inside the container.

vcpus: Option<i64>

The number of VCPUs allocated for the job.

volumes: Option<Vec<Volume>>

A list of volumes associated with the job.